diabete-backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
scikit-learn==1.7.1
numpy==1.26.4
pandas==2.1.4
joblib==1.3.2
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.1.1
